Upgraded dependencies and blog excerpts
Upgraded dependencies using "bundle update", which
introduced breaking changes to excerpts resulting
in the default behaviour being the Blogs section
displaying the full content of every blog post
rather than just the first paragraph.

Added an explicit excerpt_separator value of
<!--more--> to _config.yml that can be specified
on any line of a blog post to signal the end
of the excerpt to show as a preview.

Updated all posts without explicitly set excerpts
to either use this tag, or define a specific
excerpt, so the Blogs previews continue to work
as intended after the upgrade.